Fixed AMP CORs and Multipart form data for Request Bank
